Require Jenkins 2.414.3 or newer
Fixes #6 - Bump json jar version to latest by replacing the jar file
dependency on json-lib with a dependency on the json API library plugin.
https://plugins.jenkins.io/json-api/

https://stats.jenkins.io/pluginversions/custom-markup-formatter.html shows
that 85% of all installations of the plugin are already running Jenkins
2.414.3 or newer.  Users that are already running Jenkins 2.414.3 or newer
can easily upgrade to this plugin version without upgrading Jenkins core.

https://www.jenkins.io/security/advisory/2024-01-24/#SECURITY-3314
strongly suggests that the minimum Jenkins version should be 2.426.3
or newer due to the critical nature of that security issue.  However,
rather than risk that this pull request might be rejected because it only
covers 72% of all the installations of the plugin, I kept it at 2.414.3.

Upgrades to the most recent version of HTMLUnit so that automated tests
continue to run.

Upgrades to the most recent parent pom so that the latest capabilities
of the plugin parent pom are available.